smart-code-tutor/backend/requirements.txt
pypdf
fastapi
uvicorn
websockets
e2b_code_interpreter
python-dotenv
langchain
langchain-community
langchain-pinecone
langchain-nomic
pinecone-client
openai
# If using Groq LLM via LangChain
langchain-groq